refactor(insights): widen zoom bucket end in zoomDateRange, not per chart
Every zoomable chart funnels into insightVizDataLogic.zoomDateRange, which already knows the query's interval — so the bucket-end widening lives there instead of threading interval through useDateRangeZoom at each chart call site. The hook keeps its two-arg signature and emits bucket starts; widening is asserted through the zoomDateRange action per interval.

Generated-By: PostHog Code
Task-Id: cea09026-0a53-4a03-ae4b-dd54864a9159
