+++ products/product_analytics/frontend/insights/trends/TrendsBarChart/TrendsBarChart.tsx
@@ -363,7 +363,7 @@ export function TrendsBarChart({
     )
 
     // Time-series layouts only — the aggregated bar-value layout has categorical labels, not dates.
-    const onDateRangeZoom = useDateRangeZoom(currentPeriodResult?.days, context?.onDateRangeZoom, interval)
+    const onDateRangeZoom = useDateRangeZoom(currentPeriodResult?.days, context?.onDateRangeZoom)
 
     const renderTooltip = useCallback(
         (ctx: TooltipContext<TrendsSeriesMeta>) => {
